More files refactoring
