+++ src/run_clustering.py
@@ -4,7 +4,7 @@ from hyperopt import fmin, tpe, hp
 from hyperopt.pyll import scope
 from s2and_ext.utils import load_dataset
 from s2and_ext.clustering_models import Clusterer
-from s2and_ext.my_featurization import featurizing_function
+from s2and_ext.featurization.utils import featurizing_function
 from s2and_ext.clustering_objective import Objective
 from utils.configs import load_configurations
 from utils.mlflow import get_or_create_experiment
