+++ src/s2and_ext/featurization/utils.py
@@ -1,14 +1,16 @@
-import json, torch, numpy as np
-from typing import Callable, Dict, List, Tuple, Union
-
+import torch
+import numpy as np
+from typing import Dict, List, Tuple, Callable
 from Levenshtein import distance
-from s2and_ext.utils import load_dataset, load_signatures
 from torch.nn import functional as F
+from s2and_ext.featurization.featurizer import Featurizer
 
 
-def extract_feature(sig1 : Dict[str, dict], 
-                    sig2 : Dict[str, dict],
-                    attribute : str) -> float :
+def extract_feature(
+    sig1: Dict[str, dict], 
+    sig2: Dict[str, dict],
+    attribute : str
+) -> float:
     """
     Returns 1 if attribute is equal in sig1 and sig2 otherwise 0.
     If attribute is missing, returns np.nan
@@ -19,8 +21,10 @@ def extract_feature(sig1 : Dict[str, dict],
         return 0
     return np.nan
 
-def cosine_sim(sig1 : Dict[str, dict], 
-               sig2 : Dict[str, dict]) -> float :
+def cosine_sim(
+    sig1: Dict[str, dict], 
+    sig2: Dict[str, dict]
+) -> float:
     """
     Computes cosine similarity of paperVector field of signature.
     If field is missing, returns np.nan
@@ -35,8 +39,10 @@ def cosine_sim(sig1 : Dict[str, dict],
         return F.cosine_similarity(v1, v2, dim=0).item()
     return np.nan
 
-def name_distance(sig1 : Dict[str, dict],
-                  sig2 : Dict[str, dict]) -> float :
+def name_distance(
+    sig1: Dict[str, dict],
+    sig2: Dict[str, dict]
+) -> float:
     """
     Computes levenshtein distance of s2AuthorName attribute.
     If attribute is missing, return np.nan
@@ -47,9 +53,11 @@ def name_distance(sig1 : Dict[str, dict],
         return distance(v1,v2)
     return np.nan
 
-def jaccard(sig1 : Dict[str, dict], 
-            sig2 : Dict[str, dict],
-            attribute : str) -> float :
+def jaccard(
+    sig1: Dict[str, dict], 
+    sig2: Dict[str, dict],
+    attribute: str
+) -> float:
     """
     Computes jaccard similarity of list attribute between two signatures.
     If attribute is missing, returns np.nan
@@ -64,8 +72,10 @@ def jaccard(sig1 : Dict[str, dict],
         return 0
     return np.nan
 
-def featurizing_function(sig1 : Dict[str, dict],
-                         sig2 : Dict[str, dict]) -> float:
+def featurizing_function(
+    sig1: Dict[str, dict],
+    sig2: Dict[str, dict]
+) -> float:
     """
     Calculates the feature vector of two given signature dicts
     """
@@ -83,70 +93,13 @@ def featurizing_function(sig1 : Dict[str, dict],
     features.append(extract_feature(sig1, sig2, 'orcId'))
     return features
 
-class Featurizer():
-    '''
-    Class for creating numpy arrays containg the features for each signature pair
-    '''
-
-    def __init__(self, 
-                 dataset_name : str, 
-                 featurizing_function : Callable, 
-                 default_embeddings : bool = True,
-                 embeddings_path : Union[None, str] = None) -> None :
-        """
-        Initializes Featurizer object, loading datasets and signatures.
-        If parse_specter = true, it parses the provided specter embeddings
-        for them to be used in the featurization process.
-        """
-        self.default_embeddings = default_embeddings
-        self.dataset = load_dataset(dataset_name)
-        self.extended_signatures = load_signatures(dataset_name)
-        self.featurizing_function = featurizing_function
-
-        if not default_embeddings:
-            with open(embeddings_path) as f:
-                self.paper_ids_to_emb = json.load(f)
-
-    def featurize_pairs(self, pairs : Tuple[Dict[str, dict]]) -> Tuple[np.ndarray]:
-        '''
-        Given the list of pairs return the matrix of features and labels
-        '''  
-        X = []
-        y = []
-
-        for i, pair in enumerate(pairs):
-            # Make sure the extended dataset contains the signatures
-            #  of the pair, else do not featurize the pair
-            if pair[0] in self.extended_signatures and pair[1] in self.extended_signatures:
-                y.append(pair[2])
-                sig1 = self.extended_signatures[pair[0]]
-                sig2 = self.extended_signatures[pair[1]]
-                if not self.default_embeddings:
-                    sig1['paperVector'] = self.paper_ids_to_emb[str(sig1['paper_id'])]
-                    sig2['paperVector'] = self.paper_ids_to_emb[str(sig2['paper_id'])]
-                X.append(self.featurizing_function(sig1, sig2))
-
-        return np.asarray(X), np.asarray(y)
-
-    def get_feature_matrix(self, split : str) -> Tuple[np.ndarray]:
-        '''
-        Get dataset's featurized pairs matrix by specifying the desired split
-        '''
-        train_sig, val_sig, test_sig = self.dataset.split_cluster_signatures()
-        train_pairs, val_pairs, test_pairs = self.dataset.split_pairs(train_sig, val_sig, test_sig)
-        
-        if split == 'train':
-            return self.featurize_pairs(train_pairs)
-        elif split == 'val':
-            return self.featurize_pairs(val_pairs)
-        elif split == 'test':
-            return self.featurize_pairs(test_pairs)
-
-def get_matrices(datasets : List[str], 
-                 featurizing_function : Callable, 
-                 remove_nan : bool =True,
-                 default_embeddings: bool = True,
-                 external_emb_dir: str = None) -> Tuple[np.ndarray]:
+def get_matrices(
+    datasets : List[str], 
+    featurizing_function : Callable, 
+    remove_nan : bool =True,
+    default_embeddings: bool = True,
+    external_emb_dir: str = None
+) -> Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
     '''
     Featurize multiple datasets and return the combined matrix
     If no featurizing_function is given, the default will be used
@@ -154,23 +107,18 @@ def get_matrices(datasets : List[str],
     X_train, y_train = [], []
     X_val, y_val = [], []
     X_test, y_test = [], []
-
     for dataset_name in datasets:
-
         embeddings_path = f'{external_emb_dir}/{dataset_name}/{dataset_name}_embeddings.json'
         featurizer = Featurizer(dataset_name=dataset_name, 
                                 featurizing_function=featurizing_function,
                                 default_embeddings=default_embeddings,
                                 embeddings_path=embeddings_path)
-        
         X, y = featurizer.get_feature_matrix('train')
         X_train.append(X)
         y_train.append(y)
-
         X, y = featurizer.get_feature_matrix('val')
         X_val.append(X)
         y_val.append(y)
-        
         X, y = featurizer.get_feature_matrix('test')
         X_test.append(X)
         y_test.append(y)
@@ -178,21 +126,16 @@ def get_matrices(datasets : List[str],
 
     X_train = np.vstack(X_train)
     y_train = np.concatenate(y_train)
-
     X_val = np.vstack(X_val)
     y_val = np.concatenate(y_val)
-
     X_test = np.vstack(X_test)
     y_test = np.concatenate(y_test)
-
     # Count nan per column for train set
     print('Nan values for each feature :')
     print(np.count_nonzero(np.isnan(X_train), axis=0))
-
     # Remove nan values
     if remove_nan:
         np.nan_to_num(X_train, copy=False)
         np.nan_to_num(X_test, copy=False)
         np.nan_to_num(X_val, copy=False)
-
     return X_train, y_train, X_val, y_val, X_test, y_test
